[REFACTOR] split delivery flows into DeliveryStrategy implementations
